Add snowfall effect to About, Homepage, and Leaderboard pages
