+++ client/src/components/Contests/ContestPageLayout.jsx
@@ -2,6 +2,8 @@ import React from "react";
 import { LockOutlined, TrendingUp } from "@mui/icons-material";
 import { NavLink, Outlet } from "react-router-dom";
 import { Swords } from "lucide-react";
+import Snowfall from "react-snowfall";
+import SnowFlakes from "../Home/components/SnowFlakes";
 
 const activeLinks = [
     {
@@ -51,42 +53,48 @@ const inactiveLinks = [
 ];
 
 const ContestPageLayout = () => {
+    const [isSnowfallActive, setSnowfallActive] = React.useState(true);
     return (
-        <div className="w-11/12 mx-auto antialiased">
-            <div className="heading w-4/5 mx-auto text-center my-4">
-                <h1 className="text-white max-md:text-4xl md:text-6xl flex flex-row mx-auto justify-center">
-                    <span>All at</span>
-                    <span className="block mt-1 md:mt-6">
-                        <span className="bg-digitomize-bg mx-2 px-1 rounded-lg">one</span>
-                        place
-                    </span>
-                </h1>
+        <>
+            {isSnowfallActive && <Snowfall snowflakeCount={70} speed={[0, 0.7]} wind={[0, 0.5]} />}
+            <SnowFlakes onClick={() => setSnowfallActive(!isSnowfallActive)} position="top-left" />
+>
+            <div className="w-11/12 mx-auto antialiased">
+                <div className="heading w-4/5 mx-auto text-center my-4">
+                    <h1 className="text-white max-md:text-4xl md:text-6xl flex flex-row mx-auto justify-center">
+                        <span>All at</span>
+                        <span className="block mt-1 md:mt-6">
+                            <span className="bg-digitomize-bg mx-2 px-1 rounded-lg">one</span>
+                            place
+                        </span>
+                    </h1>
 
+                </div>
+                <div className="buttons flex gap-4 justify-center mt-8 flex-wrap">
+                    {
+                        activeLinks.map((link, idx) => {
+                            return (
+                                <NavLink key={idx} to={link.link} className={({ isActive }) => isActive ? "btn bg-custom-blue text-white" : "btn btn-outline"} >
+                                    {link.name}
+                                    {link.icon}
+                                </NavLink>
+                            );
+                        })
+                    }
+                    {
+                        inactiveLinks.map((link, idx) => {
+                            return (
+                                <NavLink aria-disabled key={idx} to={link.link} className={({ isActive }) => isActive ? "btn bg-custom-blue text-gray-700 max-sm:hidden" : "btn btn-outline btn-disabled max-sm:hidden"} style={{ color: "#7a7a7a" }}>
+                                    {link.name}
+                                    {link.icon}
+                                </NavLink>
+                            );
+                        })
+                    }
+                </div>
+                <Outlet />
             </div>
-            <div className="buttons flex gap-4 justify-center mt-8 flex-wrap">
-                {
-                    activeLinks.map((link, idx) => {
-                        return (
-                            <NavLink key={idx} to={link.link} className={({ isActive }) => isActive ? "btn bg-custom-blue text-white" : "btn btn-outline"} >
-                                {link.name}
-                                {link.icon}
-                            </NavLink>
-                        );
-                    })
-                }
-                {
-                    inactiveLinks.map((link, idx) => {
-                        return (
-                            <NavLink aria-disabled key={idx} to={link.link} className={({ isActive }) => isActive ? "btn bg-custom-blue text-gray-700 max-sm:hidden" : "btn btn-outline btn-disabled max-sm:hidden"} style={{color:"#7a7a7a"}}>
-                                {link.name}
-                                {link.icon}
-                            </NavLink>
-                        );
-                    })
-                }
-            </div>
-            <Outlet />
-        </div>
+        </>
     );
 };
